requirements.txt
Django==4.2.8
djangorestframework==3.14.0
django-filter==25.1
python-dotenv==1.0.0
pytz==2025.2
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
sqlparse==0.5.3
tzdata==2025.2
psycopg2-binary==2.9.9
redis==5.0.1
gunicorn==21.2.0
whitenoise==6.6.0
django-cors-headers==4.3.1
dj-database-url==2.1.0

# Production dependencies
setuptools==69.0.0
wheel==0.42.0
django-filter==23.5
